biome Capture: M analysis.js, M index.js, M indexed-db-adapter.js ... and 3 more files
X-Quipu-Output-Tree: 26c9cd7b7be3d59ac3a45420e619fba010d4780b

+++ snapshot/src/storage/index.js
@@ -83,8 +83,7 @@ class StorageManager {
   }
 
   deleteMessages(ids) {
-    if (this.adapter?.deleteMessages)
-      return this.adapter.deleteMessages(ids);
+    if (this.adapter?.deleteMessages) return this.adapter.deleteMessages(ids);
     return Promise.resolve();
   }
